+++ includes/features/class-events.php
@@ -660,7 +660,7 @@ class Events extends \WP_List_Table {
 					} else {
 						$local = false;
 					}
-					if ( Role::SUPER_ADMIN === Role::admin_type() || Role::SINGLE_ADMIN === Role::admin_type() || ( Role::LOCAL_ADMIN === Role::admin_type() && $local ) ) {
+					if ( Role::SUPER_ADMIN === Role::admin_type() || Role::SINGLE_ADMIN === Role::admin_type() || ( Role::LOCAL_ADMIN === Role::admin_type() && $local ) || Role::override_privileges() ) {
 						$log = [
 							'name'    => $logger['name'],
 							'running' => $logger['running'],